Move MutuallyExclusiveOptions to the components package
